+++ AllReadyApp/Web-App/AllReady/Models/Activity.cs
@@ -5,42 +5,41 @@ using System.ComponentModel.DataAnnotations.Schema;
 
 namespace AllReady.Models
 {
-    public class Activity
-    {
+  public class Activity
+  {
 
-        [Key, DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public int Id { get; set; }
+    public int Id { get; set; }
 
-        [Display(Name = "Tenant")]
-        public int TenantId { get; set; }
+    [Display(Name = "Tenant")]
+    public int TenantId { get; set; }
 
-        public Tenant Tenant { get; set; }
+    public Tenant Tenant { get; set; }
 
-        [Display(Name = "Campaign")]
-        public int CampaignId { get; set; }
+    [Display(Name = "Campaign")]
+    public int CampaignId { get; set; }
 
-        public Campaign Campaign { get; set; }
+    public Campaign Campaign { get; set; }
 
-        public string Name { get; set; }
+    public string Name { get; set; }
 
-        public string Description { get; set; }
+    public string Description { get; set; }
 
-        [Display(Name = "Start date")]
-        public DateTime StartDateTimeUtc { get; set; }
+    [Display(Name = "Start date")]
+    public DateTime StartDateTimeUtc { get; set; }
 
-        [Display(Name = "End date")]
-        public DateTime EndDateTimeUtc { get; set; }
+    [Display(Name = "End date")]
+    public DateTime EndDateTimeUtc { get; set; }
 
-        public Location Location { get; set; }
+    public Location Location { get; set; }
 
-        public List<AllReadyTask> Tasks { get; set; }
+    public List<AllReadyTask> Tasks { get; set; }
 
-        public List<ActivitySignup> UsersSignedUp { get; set; }
+    public List<ActivitySignup> UsersSignedUp { get; set; }
 
-        public ApplicationUser Organizer { get; set; }
+    public ApplicationUser Organizer { get; set; }
 
-        [Display(Name="Image")]
-        public string ImageUrl { get; set; }
+    [Display(Name = "Image")]
+    public string ImageUrl { get; set; }
 
-    }
+  }
 }
