+++ AllReadyApp/Web-App/AllReady/Models/AllReadyTask.cs
@@ -5,25 +5,15 @@ using System.ComponentModel.DataAnnotations.Schema;
 
 namespace AllReady.Models
 {
-    public class AllReadyTask
-    {
-
-        [Key, DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public int Id { get; set; }
-
-        public string Name { get; set; }
-
-        public string Description { get; set; }
-
-        public Tenant Tenant { get; set; }
-
-        public Activity Activity { get; set; }
-
-        public DateTime? StartDateTimeUtc { get; set; }
-
-        public DateTime? EndDateTimeUtc { get; set; }
-
-        public List<TaskUsers> AssignedVolunteers { get; set; }
-
-    }
+  public class AllReadyTask
+  {
+    public int Id { get; set; }
+    public string Name { get; set; }
+    public string Description { get; set; }
+    public Tenant Tenant { get; set; }
+    public Activity Activity { get; set; }
+    public DateTime? StartDateTimeUtc { get; set; }
+    public DateTime? EndDateTimeUtc { get; set; }
+    public List<TaskUsers> AssignedVolunteers { get; set; }
+  }
 }
