+++ AllReadyApp/Web-App/AllReady/Models/Campaign.cs
@@ -7,20 +7,12 @@ namespace AllReady.Models
 {
     public class Campaign
     {
-
-        [Key, DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         public int Id { get; set; }
-
-        [Required]
         public string Name { get; set; }
-
         public string Description { get; set; }
-
         [Display(Name = "Managing Tenant")]
         public int ManagingTenantId { get; set; }
-
         public Tenant ManagingTenant { get; set; }
-
         [Display(Name = "Image URL")]
         public string ImageUrl { get; set; }
