+++ AllReadyApp/Web-App/AllReady/Models/CampaignSponsors.cs
@@ -3,15 +3,11 @@ using System.ComponentModel.DataAnnotations.Schema;
 
 namespace AllReady.Models
 {
-    public class CampaignSponsors
-    {
+  public class CampaignSponsors
+  {
+    public int Id { get; set; }
+    public Campaign Campaign { get; set; }
+    public Tenant Tenant { get; set; }
 
-        [Key, DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public int Id { get; set; }
-
-        public Campaign Campaign { get; set; }
-
-        public Tenant Tenant { get; set; }
-
-    }
+  }
 }
