+++ AllReadyApp/Web-App/AllReady/Models/PostalCodeGeoCoordinate.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace AllReady.Models
+{
+  public class PostalCodeGeoCoordinate
+  {
+    public double Latitude { get; set; }
+    public double Longitude { get; set; }
+  }
+}
