+++ AllReadyApp/Web-App/AllReady/Models/Resource.cs
@@ -7,16 +7,15 @@ using System.Threading.Tasks;
 
 namespace AllReady.Models
 {
-    public class Resource
-    {
-        [Key, DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public int Id { get; set; }
-        public string Name { get; set; }
-        public string Description { get; set; }
-        public DateTime PublishDateBegin { get; set; }
-        public DateTime PublishDateEnd { get; set; }
-        public string MediaUrl { get; set; }
-        public string ResourceUrl { get; set; }
-        public string CategoryTag { get; set; }
-    }
+  public class Resource
+  {
+    public int Id { get; set; }
+    public string Name { get; set; }
+    public string Description { get; set; }
+    public DateTime PublishDateBegin { get; set; }
+    public DateTime PublishDateEnd { get; set; }
+    public string MediaUrl { get; set; }
+    public string ResourceUrl { get; set; }
+    public string CategoryTag { get; set; }
+  }
 }
