+++ AllReadyApp/Web-App/AllReady/Models/Security.cs
@@ -6,17 +6,17 @@ using System.Threading.Tasks;
 namespace AllReady.Models
 {
 
-    public enum TenantSecurity
-    {
-        Volunteer = 1,
-        Employee = 2,
-        Admin = 3
-    }
+  public enum TenantSecurity
+  {
+    Volunteer = 1,
+    Employee = 2,
+    Admin = 3
+  }
 
-    public enum UserType
-    {
-        BasicUser,
-        TenantAdmin,
-        SiteAdmin
-    }
+  public enum UserType
+  {
+    BasicUser,
+    TenantAdmin,
+    SiteAdmin
+  }
 }
