+++ AllReadyApp/Web-App/AllReady/Models/TaskUsers.cs
@@ -4,21 +4,13 @@ using System.ComponentModel.DataAnnotations.Schema;
 
 namespace AllReady.Models
 {
-    public class TaskUsers
-    {
-
-        [Key, DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public int Id { get; set; }
-
-        public AllReadyTask Task { get; set; }
-
-        public ApplicationUser User { get; set; }
-
-        public DateTime StatusDateTimeUtc { get; set; } = DateTime.UtcNow;
-
-        public string Status { get; set; }
-
-        public string StatusDescription { get; set; }
-
-    }
+  public class TaskUsers
+  {
+    public int Id { get; set; }
+    public AllReadyTask Task { get; set; }
+    public ApplicationUser User { get; set; }
+    public DateTime StatusDateTimeUtc { get; set; } = DateTime.UtcNow;
+    public string Status { get; set; }
+    public string StatusDescription { get; set; }
+  }
 }
